Move types and styles in different files
