+++ packages/core/src/js/feedback/FeedbackFormScreen.tsx
@@ -1,105 +1,14 @@
 import { captureFeedback } from '@sentry/core';
 import type { SendFeedbackParams } from '@sentry/types';
 import * as React from 'react';
-import type { KeyboardTypeOptions, TextStyle,ViewStyle } from 'react-native';
-import { Alert, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import type { KeyboardTypeOptions } from 'react-native';
+import { Alert, Text, TextInput, TouchableOpacity, View } from 'react-native';
 
-interface FeedbackFormScreenProps {
-  closeScreen: () => void;
-  text: FeedbackFormText;
-  styles?: FeedbackFormScreenStyles;
-}
-
-interface FeedbackFormText {
-  formTitle?: string;
-  namePlaceholder?: string;
-  emailPlaceholder?: string;
-  descriptionPlaceholder?: string;
-  addAttachmentButton?: string;
-  submitButton?: string;
-  cancelButton?: string;
-  formError?: string;
-}
-
-interface FeedbackFormScreenStyles {
-  container?: ViewStyle;
-  title?: TextStyle;
-  input?: TextStyle;
-  textArea?: ViewStyle;
-  screenshotButton?: ViewStyle;
-  screenshotText?: TextStyle;
-  submitButton?: ViewStyle;
-  submitText?: TextStyle;
-  cancelButton?: ViewStyle;
-  cancelText?: TextStyle;
-}
-
-interface FeedbackFormScreenState {
-  name: string;
-  email: string;
-  description: string;
-}
-
-const defaultStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: '#fff',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  input: {
-    height: 50,
-    borderColor: '#ccc',
-    borderWidth: 1,
-    borderRadius: 5,
-    paddingHorizontal: 10,
-    marginBottom: 15,
-    fontSize: 16,
-  },
-  textArea: {
-    height: 100,
-    textAlignVertical: 'top',
-  },
-  screenshotButton: {
-    backgroundColor: '#eee',
-    padding: 15,
-    borderRadius: 5,
-    marginBottom: 20,
-    alignItems: 'center',
-  },
-  screenshotText: {
-    color: '#333',
-    fontSize: 16,
-  },
-  submitButton: {
-    backgroundColor: '#6a1b9a',
-    paddingVertical: 15,
-    borderRadius: 5,
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  submitText: {
-    color: '#fff',
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  cancelButton: {
-    paddingVertical: 15,
-    alignItems: 'center',
-  },
-  cancelText: {
-    color: '#6a1b9a',
-    fontSize: 16,
-  },
-});
+import defaultStyles from './FeedbackFormScreen.styles';
+import type { FeedbackFormScreenProps, FeedbackFormScreenState } from './FeedbackFormScreen.types';
 
 /**
- *
+ * Implements a feedback form screen that sends feedback to Sentry using Sentry.captureFeedback.
  */
 export class FeedbackFormScreen extends React.Component<FeedbackFormScreenProps, FeedbackFormScreenState> {
   public constructor(props: FeedbackFormScreenProps) {
@@ -137,7 +46,7 @@ export class FeedbackFormScreen extends React.Component<FeedbackFormScreenProps,
   };
 
   /**
-   *
+   * Renders the feedback form screen.
    */
   public render(): React.ReactNode {
     const { closeScreen, text, styles } = this.props;
